refactor(json): move go.starlark.net/starlarkjson -> go.starlark.net/lib/json
and keep an alias package in place for backwards compatibility
